requirements.txt
pdfplumber>=0.10.0
docling>=2.0.0
PyMuPDF>=1.24.0
ocrmac>=1.0.0
streamlit>=1.30.0
openai
python-dotenv
